Create own file for SearchUtils
pickSupportedFilters is now under SearchUtils as it's
only useful with searches.
