+++ app/utils/__tests__/DataUtils.spec.js
@@ -10,7 +10,6 @@ import {
   getOpeningHours,
   getPeopleCapacityString,
   humanizeMainType,
-  pickSupportedFilters,
 } from 'utils/DataUtils';
 
 describe('Utils: DataUtils', () => {
@@ -275,22 +274,6 @@ describe('Utils: DataUtils', () => {
       const expected = PURPOSE_MAIN_TYPES[validType];
 
       expect(humanizeMainType(mainType)).to.equal(expected);
-    });
-  });
-
-  describe('pickSupportedFilters', () => {
-    it('should only return supported filters', () => {
-      const filters = {
-        purpose: 'some-purpose',
-        search: 'search-query',
-        unsupported: 'invalid',
-      };
-      const expected = {
-        purpose: 'some-purpose',
-        search: 'search-query',
-      };
-
-      expect(pickSupportedFilters(filters)).to.deep.equal(expected);
     });
   });
 });
